Extract reusable filtered/paginated list shell for dashboard pages
- Add useListFilters() hook: nuqs-backed URL state for the standard
  search/tags/model/lookback filter set plus page-specific enum keys
  (extra), with isFiltered + clearFilters in one place.
- Add FilteredListShell: composes FilterBar with the result count
  ("n of m results"), loading/error states, and the two distinct empty
  states ("no data" vs "no matches for your filters").
- Add ListPagination + usePagedList: shared page-based pagination,
  prop shape ready for server-driven page/total later.
- Split citations-display.tsx (1,391 lines) into focused subcomponents
  under components/citations/; the composer keys section visibility off
  the unfiltered data so in-card filters can't hide a section (#322).
- Convert the citations and visibility/prompts pages to the hook +
  shell; share-of-voice and the visibility bar move to useListFilters;
  remove usePageFilters/usePageFilterSetters.
