+++ .changeset/filter-empty-states.md
@@ -1,0 +1,5 @@
+---
+"@workspace/web": patch
+---
+
+Made dashboard filter empty states consistent, and Clear filters now also resets the model filter.
