+++ apps/web/src/routes/_authed/app/$brand/citations.tsx
@@ -8,14 +8,15 @@ import { useQueryClient } from "@tanstack/react-query";
 import { getAppName, getBrandName, buildTitle } from "@/lib/route-head";
 import { Card, CardContent, CardHeader } from "@workspace/ui/components/card";
 import { Skeleton } from "@workspace/ui/components/skeleton";
-import { Button } from "@workspace/ui/components/button";
 import { useCitations } from "@/hooks/use-citations";
 import { useBrand, brandKeys } from "@/hooks/use-brands";
+import { useListFilters } from "@/hooks/use-list-filters";
 import { dashboardKeys } from "@/hooks/use-dashboard-summary";
 import { CitationsDisplay } from "@/components/citations-display";
+import { FilteredListShell } from "@/components/filtered-list-shell";
 import { getDaysFromLookback } from "@/lib/chart-utils";
-import { PageHeader, FilterSection } from "@/components/page-header";
-import { FilterBar, getAvailableModels, usePageFilters, usePageFilterSetters } from "@/components/filter-bar";
+import { PageHeader } from "@/components/page-header";
+import { getAvailableModels, ALL_MODELS_VALUE } from "@/components/filter-bar";
 
 
 export const Route = createFileRoute("/_authed/app/$brand/citations")({
@@ -36,15 +37,14 @@ function CitationsPage() {
 	const { brand: brandId } = Route.useParams();
 	const queryClient = useQueryClient();
 
-	const { selectedModel, selectedLookback, selectedTags } = usePageFilters();
-	const { clearFilters } = usePageFilterSetters();
-	const days = getDaysFromLookback(selectedLookback);
+	const filters = useListFilters();
+	const days = getDaysFromLookback(filters.lookback);
 
 	const { brand } = useBrand(brandId);
 	const availableModels = getAvailableModels(brand?.effectiveModels ?? []);
 
 	// Get citation data with tag and model filter
-	const modelParam = selectedModel === "all" ? undefined : selectedModel;
+	const modelParam = filters.model === ALL_MODELS_VALUE ? undefined : filters.model;
 	const {
 		citations: citationData,
 		isLoading,
@@ -52,11 +52,9 @@ function CitationsPage() {
 		revalidate: revalidateCitations,
 	} = useCitations(brandId, {
 		days,
-		tags: selectedTags.length > 0 ? selectedTags : undefined,
+		tags: filters.tags.length > 0 ? filters.tags : undefined,
 		model: modelParam,
 	});
-
-	const availableTags = citationData?.availableTags || [];
 
 	const infoContent = (
 		<>
@@ -74,71 +72,6 @@ function CitationsPage() {
 	);
 
 	const showFullSkeleton = isLoading && !citationData;
-	let content: React.ReactNode;
-	if (showFullSkeleton) {
-		content = (
-			<Card>
-				<CardHeader>
-					<Skeleton className="h-6 w-48" />
-				</CardHeader>
-				<CardContent>
-					<div className="space-y-4">
-						<Skeleton className="h-4 w-3/4" />
-						<Skeleton className="h-4 w-1/2" />
-						<Skeleton className="h-4 w-2/3" />
-					</div>
-				</CardContent>
-			</Card>
-		);
-	} else if (isError || !citationData) {
-		content = (
-			<Card>
-				<CardContent className="pt-6">
-					<div className="text-red-600 text-sm bg-red-50 p-3 rounded-md">
-						Failed to load citation data. Please try again.
-					</div>
-				</CardContent>
-			</Card>
-		);
-	} else if (citationData.totalCitations === 0) {
-		content = (
-			<Card>
-				<CardContent className="pt-6">
-					<div className="text-muted-foreground text-center py-8">
-						{selectedTags.length > 0 || selectedModel !== "all" ? (
-							<>
-								<p className="mb-2">No citations found for the selected filters.</p>
-								<p className="text-sm mb-4">Try adjusting your filters or time period.</p>
-								<Button variant="outline" size="sm" onClick={clearFilters} className="cursor-pointer">
-									Clear filters
-								</Button>
-							</>
-						) : (
-							"No citations found. Citations are only available from prompts evaluated with web search enabled."
-						)}
-					</div>
-				</CardContent>
-			</Card>
-		);
-	} else {
-		content = (
-			<CitationsDisplay
-				citationData={citationData}
-				brandId={brandId}
-				brandName={brand?.name}
-				showStats={true}
-				maxDomains={10}
-				maxUrls={20}
-				days={days}
-				onCompetitorAdded={() => {
-					revalidateCitations();
-					queryClient.invalidateQueries({ queryKey: dashboardKeys.all });
-					queryClient.invalidateQueries({ queryKey: brandKeys.competitors(brandId) });
-					queryClient.invalidateQueries({ queryKey: brandKeys.detail(brandId) });
-				}}
-			/>
-		);
-	}
 
 	return (
 		<PageHeader
@@ -146,15 +79,68 @@ function CitationsPage() {
 			subtitle="See which sources LLMs cite when responding to your prompts."
 			infoContent={infoContent}
 		>
-			<FilterSection>
-				<FilterBar
-					availableTags={availableTags}
-					availableModels={availableModels}
-					showSearch={false}
-					showModelSelector
-				/>
-			</FilterSection>
-			<div className="space-y-6">{content}</div>
+			<FilteredListShell
+				filters={filters}
+				availableTags={citationData?.availableTags || []}
+				availableModels={availableModels}
+				showSearch={false}
+				showModelSelector
+				isLoading={showFullSkeleton}
+				loadingState={
+					<Card>
+						<CardHeader>
+							<Skeleton className="h-6 w-48" />
+						</CardHeader>
+						<CardContent>
+							<div className="space-y-4">
+								<Skeleton className="h-4 w-3/4" />
+								<Skeleton className="h-4 w-1/2" />
+								<Skeleton className="h-4 w-2/3" />
+							</div>
+						</CardContent>
+					</Card>
+				}
+				isError={Boolean(isError) || !citationData}
+				errorState={
+					<Card>
+						<CardContent className="pt-6">
+							<div className="text-red-600 text-sm bg-red-50 p-3 rounded-md">
+								Failed to load citation data. Please try again.
+							</div>
+						</CardContent>
+					</Card>
+				}
+				totalCount={citationData?.totalCitations}
+				noMatchesTitle="No citations found for the selected filters."
+				noMatchesDescription="Try adjusting your filters or time period."
+				emptyState={
+					<Card>
+						<CardContent className="pt-6">
+							<div className="text-muted-foreground text-center py-8">
+								No citations found. Citations are only available from prompts evaluated with web search enabled.
+							</div>
+						</CardContent>
+					</Card>
+				}
+			>
+				{citationData && (
+					<CitationsDisplay
+						citationData={citationData}
+						brandId={brandId}
+						brandName={brand?.name}
+						showStats={true}
+						maxDomains={10}
+						maxUrls={20}
+						days={days}
+						onCompetitorAdded={() => {
+							revalidateCitations();
+							queryClient.invalidateQueries({ queryKey: dashboardKeys.all });
+							queryClient.invalidateQueries({ queryKey: brandKeys.competitors(brandId) });
+							queryClient.invalidateQueries({ queryKey: brandKeys.detail(brandId) });
+						}}
+					/>
+				)}
+			</FilteredListShell>
 		</PageHeader>
 	);
 }
